updated all core go file packages based on Mono repo, updated makefile for core services, added cmd folders for core services
