+++ core/metadata/VERSION
@@ -1,0 +1,1 @@
+0.5.0
